Separated initMap to own separate file
