+++ client/src/layouts/ProfileSummary/ProfileSummary.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import styled from "styled-components";
 import PropTypes from "prop-types";
 import ProfileSummaryHeader from "./ProfileSummaryHeader";
